Basic save-data-to-file/load-data-from-file functionality
This was helpful when I was trying to debug without a network
connection.

Saving this to a branch so it's not lost, but I think this can serve as
a basis for when I want to start playing about with long-running
staggered API walks, for very large repos that cannot be aggregated in
one pass. The contributions can go into one file, the non-code needs to be
staggered, or some sort of code-drive staggered something with state
saving and other such fun.